+++ src/main/java/com/bbangle/bbangle/boardstatistic/service/BoardStatisticService.java
@@ -2,12 +2,15 @@ package com.bbangle.bbangle.boardstatistic.service;
 
 import com.bbangle.bbangle.board.domain.Board;
 import com.bbangle.bbangle.board.repository.BoardRepository;
+import com.bbangle.bbangle.exception.BbangleErrorCode;
+import com.bbangle.bbangle.exception.BbangleException;
 import com.bbangle.bbangle.boardstatistic.domain.BoardStatistic;
 import com.bbangle.bbangle.boardstatistic.repository.BoardStatisticRepository;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -30,6 +33,14 @@ public class BoardStatisticService {
                 .build())
             .forEach(boardStatistics::add);
         boardStatisticRepository.saveAll(boardStatistics);
+    }
+
+    @Transactional
+    public void updateRankingScore(Long boardId, Double updatingScore) {
+        BoardStatistic boardStatistic = boardStatisticRepository.findByBoardId(boardId)
+            .orElseThrow(
+                () -> new BbangleException(BbangleErrorCode.RANKING_NOT_FOUND));
+        boardStatistic.updateBasicScore(updatingScore);
     }
 
 }
