--- src/main/java/com/bbangle/bbangle/boardstatistic/service/RankingService.java
@@ -1,46 +1,0 @@
-package com.bbangle.bbangle.boardstatistic.service;
-
-import com.bbangle.bbangle.board.domain.Board;
-import com.bbangle.bbangle.board.repository.BoardRepository;
-import com.bbangle.bbangle.exception.BbangleErrorCode;
-import com.bbangle.bbangle.exception.BbangleException;
-import com.bbangle.bbangle.boardstatistic.domain.BoardStatistic;
-import com.bbangle.bbangle.boardstatistic.repository.BoardStatisticRepository;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class RankingService {
-
-    private final BoardStatisticRepository boardStatisticRepository;
-    private final BoardRepository boardRepository;
-
-    public void updatingNonRankedBoards() {
-        List<Board> unRankedBoards = boardRepository.checkingNullRanking();
-        List<BoardStatistic> rankings = new ArrayList<>();
-        unRankedBoards.stream()
-            .map(board -> BoardStatistic.builder()
-                .boardId(board.getId())
-                .basicScore(0.0)
-                .boardReviewGrade(0.0)
-                .boardReviewCount(0)
-                .boardWishCount(0)
-                .boardViewCount(0)
-                .build())
-            .forEach(rankings::add);
-        boardStatisticRepository.saveAll(rankings);
-    }
-
-    @Transactional
-    public void updateRankingScore(Long boardId, Double updatingScore) {
-        BoardStatistic boardStatistic = boardStatisticRepository.findByBoardId(boardId)
-            .orElseThrow(
-                () -> new BbangleException(BbangleErrorCode.RANKING_NOT_FOUND));
-        boardStatistic.updateBasicScore(updatingScore);
-    }
-
-}
